tasks: set-timezone: Add packages dependency.
Installs tzdata if a timezone is set. Removed phoenix as the default value.

+++ tasks/image_configuration/set-timezone/set-timezone.sh
@@ -1,4 +1,6 @@
-#!/bin/bash
+#!/bin/bash -e
 
-rm /etc/localtime
+if [ -e /etc/localtime ]; then
+        rm /etc/localtime
+fi
 ln -s /usr/share/zoneinfo/${CONFIG_DS_TIMEZONE} /etc/localtime
